user-service/requirements.txt
alembic
pydantic-settings
pydantic[email]
fastapi
uvicorn
psycopg2-binary
sqlalchemy
bcrypt
passlib
python-jose
aiosmtplib
pytest
pytest-asyncio
slowapi
httpx
python-multipart
pytest-httpx
prometheus-fastapi-instrumentator
opentelemetry-instrumentation-fastapi
opentelemetry-sdk 
opentelemetry-exporter-otlp